refactor: feedscreen upload button moved to seperate componnent
